refactor(#742): extract compileCallExpression early-guards into calls-guards.ts
First incremental, behaviour-preserving step toward decomposing the
~9,400-line compileCallExpression. Pulls the self-contained early-guard
prelude into a new src/codegen/expressions/calls-guards.ts, each as a
(ctx, fctx, expr) => InnerResult | undefined handler (undefined = not my
case, caller continues dispatch):

- tryNamespaceNonCallable  Math()/JSON()/Reflect()/Atomics()/Proxy() throw (#1732/#2180)
- tryJsxRuntimeCall        _jsx/_jsxs/_jsxDEV runtime intercept (#1540)
- tryRegExpConstructorCall RegExp(p, f) without `new`
- tryObjectCoercionCall    Object(x) ToObject coercion (#1129/#1568)

compileCallExpression: 9,437 -> 9,242 lines. Behaviour-preserving — a
WAT-hash oracle over 25 call-heavy programs is byte-identical before/after;
full typecheck + prettier + biome clean.

Proceeds incrementally off current origin/main (the previous extraction
attempt was reverted for doing too much at once / branching wrong). Issue
stays in-progress; follow-up PRs continue the same pattern.

Tests: tests/issue-742.test.ts — wasm≡JS for the extracted guards.

✓
